+++ resources/views/livewire/cms/cms-footer.blade.php
@@ -1,0 +1,6 @@
+<footer class="py-3 mt-auto footer bg-body-tertiary">
+    <div class="container">
+        <span class="text-body-secondary"><a href="https://arvid.nl" class="text-secondary" target="_blank">Darvis</a>
+            {{ date('d-m-Y') }}</span>
+    </div>
+</footer>