+++ src/Livewire/Cms/CmsFooter.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Manta\FluxCMS\Livewire\Cms;
+
+use Livewire\Component;
+
+class CmsFooter extends Component
+{
+    public function render()
+    {
+        return view('manta-cms::livewire.cms.cms-footer');
+    }
+}